requirements.txt
streamlit>=1.28.0
mtcnn>=0.1.1
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
werkzeug>=3.0.0